+++ management-core/src/main/java/com/ums/management/core/service/impl/OrganizationServiceImpl.java
@@ -7,9 +7,10 @@ import java.util.Map;
 import com.ums.management.core.dao.OrganizationMapper;
 import com.ums.management.core.dao.UserOrgMapper;
 import com.ums.management.core.model.Organization;
-import com.ums.management.core.model.User;
 import com.ums.management.core.model.UserOrg;
 import com.ums.management.core.service.IOrganizationService;
+import com.ums.management.core.utility.RoleMatrix;
+import com.ums.management.core.view.model.ServiceResult;
 import com.ums.management.core.view.model.UserVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -67,13 +68,21 @@ public class OrganizationServiceImpl implements IOrganizationService {
     }
 
     @Override
-    public void deleteById(int id) {
+    public ServiceResult<Void> deleteById(UserVO editor, int id) {
+        if(!RoleMatrix.hasEnoughPower(editor.getRole())) {
+            return ServiceResult.NO_PERMISSION;
+        }
         _dao.deleteByPrimaryKey(id);
+        return ServiceResult.SUCCESS;
     }
 
     @Override
     @Transactional
-    public void create(UserVO editor, Organization organization) {
+    public ServiceResult<Void> create(UserVO editor, Organization organization) {
+        if(!RoleMatrix.hasEnoughPower(editor.getRole())) {
+            return ServiceResult.NO_PERMISSION;
+        }
+
         _dao.insert(organization);
         if (!editor.getRole().isAdmin()) {
             UserOrg uo = new UserOrg();
@@ -81,10 +90,15 @@ public class OrganizationServiceImpl implements IOrganizationService {
             uo.setOrgId(organization.getId());
             _uoDao.insert(uo);
         }
+        return ServiceResult.SUCCESS;
     }
 
     @Override
-    public void update(Organization organization) {
+    public ServiceResult<Void> update(UserVO editor, Organization organization) {
+        if(!RoleMatrix.hasEnoughPower(editor.getRole())) {
+            return ServiceResult.NO_PERMISSION;
+        }
         _dao.updateByPrimaryKey(organization);
+        return ServiceResult.SUCCESS;
     }
 }
